factorize wrap_iframe_in_div_tag in utils
