+++ lacommunaute/forum/forms.py
@@ -1,5 +1,3 @@
-import re
-
 from django import forms
 from django.conf import settings
 from django.forms import CharField, CheckboxSelectMultiple, ModelMultipleChoiceField
@@ -7,19 +5,7 @@ from taggit.models import Tag
 
 from lacommunaute.forum.models import Forum
 from lacommunaute.partner.models import Partner
-
-
-def wrap_iframe_in_div_tag(text):
-    # iframe tags must be wrapped in a div tag to be displayed correctly
-    # add div tag if not present
-
-    iframe_regex = r"((<div>)?<iframe.*?</iframe>(</div>)?)"
-
-    for match, starts_with, ends_with in re.findall(iframe_regex, text, re.DOTALL):
-        if not starts_with and not ends_with:
-            text = text.replace(match, f"<div>{match}</div>")
-
-    return text
+from lacommunaute.utils.html import wrap_iframe_in_div_tag
 
 
 class ForumForm(forms.ModelForm):
